+++ python/tvm/relax/backend/metal/coreml.py
@@ -24,7 +24,7 @@ import tvm_ffi
 
 import tvm
 from tvm.contrib import coreml_runtime
-from tvm.contrib.xcode import compile_coreml
+from tvm.support.xcode import compile_coreml
 from tvm.relax import transform
 from tvm.relax.dpl.pattern import is_op, wildcard
 from tvm.relax.expr import (
