+++ tests/python/codegen/test_target_codegen_llvm.py
@@ -23,7 +23,7 @@ import pytest
 
 import tvm
 import tvm.testing
-from tvm.contrib import clang, utils
+from tvm.support import clang, utils
 from tvm.script import ir as I
 from tvm.script import tirx as T
 from tvm.target.codegen import llvm_get_intrinsic_name, llvm_lookup_intrinsic_id
